long

+++ OTB/src/main/java/com/registro/usuarios/modelo/Articulo.java
@@ -29,7 +29,7 @@ public class Articulo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private Integer articuloId;
+    private Long articuloId;
     
     @Column
     private String titulo;
